Adds the concept of actions
Actions are named user interactions. They have handlers, which are
invoked when the interaction occurs (such as on a button click). They
may have 'query' functions, which say whether or not the action is
active on the current selection. More details in actions.js.

This makes the toolbar class more inheritable, but the change is not
wholly backward compatible. The next commit will include updated
examples showing how to build a toolbar using actions.