Editor#whenReady delays function until editor is ready
This gets around the problem whereby the iFrame onload event does not
occur immediately in Gecko browsers. This was causing the editor
'attach' test to fail in those browsers. This change allows the test to
pass, and the pass effectively tests this change. The whenReady approach
could be taken further -- it can replace the initializer block in
Editor#attach, and even the callback argument to iFrame#create, with a
simpler evented approach.

(It also lets me register key events on the editor at any stage, which
 is very helpful for shortcuts...)